feat: add flags to get total count of unread notifications (#8)
The count is saved to a file named `unread_count` in the directory
specified by the `$GH_ND_DATA_DIR` environment variable, which defaults
to: `~/.local/state/gh-notify-desktop/`

The `-c` flag respects GitHub's Last-Modified header, meaning it will
only fetch the count when the user receives new notifications. This
means the actual count can be less than the saved number if you've read
notifications since you received a new one. However, the actual count
will never be larger.

The `-C` flag ignores GitHub's polling headers when fetching the count,
so it will always be correct. The flag also skips displaying any new
desktop notifications. The count is printed to stdout in addition to
being saved in the `unread_count` file.